Fix Xcode 11.3.1 detection when project included in an Xcode workspace
When using a Tulsi-generated project by itself in Xcode 11.3.1, Xcode sets
XCODE_VERSION_ACTUAL to 1131, which Tulsi parses correctly to get the version
11.3.1.

But when used in an Xcode workspace, Xcode sets XCODE_VERSION_ACTUAL to 1130,
leading to Tulsi parsing 11.3 as the version, and later attempts to use
a non-existing version of Xcode to build, which results in an error.

Fix by ignoring XCODE_VERSION_ACTUAL because it's unreliable; instead,
parse Xcode's version.plist to read the 'CFBundleShortVersionString' entry,
which seems to return correct results.

Caveat: this implementation uses plistlib, which has a slightly different API
in python2 and python3. This change uses the python2 API, and will
need to be updated to work with python3 in the future.

Change-Id: I0f4875c34a0e59fe66fe1624676936e9ef2325ba